updater refactor: separated filename extraction from link from save_file method refactor: created site_files file to handle saving/loading of data files - moved set_filename_from_link to site_files
