+++ modules/updater/data_processing/site_files.py
@@ -1,0 +1,17 @@
+import json
+import os
+
+from modules.websites import search_links
+
+
+def set_filename_from_link(link: str, extension: str = "html") -> str:
+    """Generate filename based on link key in search_links dict"""
+    link_name = next((key for key, value in search_links.items() if value == link), link)
+    return os.path.join("modules\\updater\\sites\\HTML", f"{link_name}.{extension}")
+
+
+def load_json(filename) -> dict:
+    with open(filename, "r", encoding="utf-8") as f:
+        file = f.read()
+        data = json.loads(file)
+        return data
